refactor(search): remove search command
BREAKING CHANGE: remove search command such that urban dictionary cli is triggered by udict only
